// iOSDriver/src/generated/contractBundle.ts
export const CONTRACT_BUNDLE_METADATA = {
  "protocolVersion": "1",
  "contractVersion": "1.0.0",
  "generatorVersion": "1",
  "contractHash": "sha256:f1ca3ffa43d7418af9be345ce7daf5dcb5906e1ab6c6a764ee2faa151421d84b"
} as const;